refactor(cdk-explorer): centralize private toolkit-lib imports in api-private
Route all relative deep imports into toolkit-lib internals through a single
lib/api-private.ts barrel that carries the sole import/no-relative-packages
disable, instead of scattering inline eslint-disable comments across
consumers. Mirrors packages/aws-cdk/lib/api-private.ts.

Covers findCreationStackTrace (source-tracing), plus WATCH_EXCLUDE_DEFAULTS
and createIgnoreMatcher used by the LSP server.
