+++ packages/@aws-cdk/cdk-explorer/lib/api-private.ts
@@ -1,0 +1,4 @@
+/* eslint-disable import/no-relative-packages */
+export { findCreationStackTrace } from '../../toolkit-lib/lib/api/source-tracing/private/stack-source-tracing';
+export { WATCH_EXCLUDE_DEFAULTS } from '../../toolkit-lib/lib/actions/watch/private/helpers';
+export { createIgnoreMatcher } from '../../toolkit-lib/lib/util/glob-matcher';
