+++ packages/@aws-cdk/cdk-explorer/lib/core/assembly-reader.ts
@@ -2,9 +2,8 @@ import * as fs from 'fs';
 import * as path from 'path';
 import { buildConstructTreeAsync, CloudAssembly, MANIFEST_FILE, type ConstructTreeNode } from '@aws-cdk/cloud-assembly-api';
 import { VALIDATION_REPORT_FILE, type PolicyValidationReportJson } from '@aws-cdk/cloud-assembly-schema';
-/* eslint-disable import/no-relative-packages */
+import { findCreationStackTrace } from '../api-private';
 import { SourceMapResolver, isWithinRoot, type SourceLocation } from './source-resolver';
-import { findCreationStackTrace } from '../../../toolkit-lib/lib/api/source-tracing/private/stack-source-tracing';
 
 /**
  * A construct from the cloud assembly, decorated with the user source location
